+++ shared/src/commonMain/kotlin/data/credentials/AgeVerificationCredentialClaimDefinition.kt
@@ -1,0 +1,16 @@
+package data.credentials
+
+enum class AgeVerificationCredentialClaimDefinition {
+    AGE_OVER_12,
+    AGE_OVER_13,
+    AGE_OVER_14,
+    AGE_OVER_16,
+    AGE_OVER_18,
+    AGE_OVER_21,
+    AGE_OVER_25,
+    AGE_OVER_60,
+    AGE_OVER_62,
+    AGE_OVER_65,
+    AGE_OVER_68,
+    ;
+}