+++ shared/src/commonMain/kotlin/ui/composables/credentials/AgeVerificationCredentialSummaryCardContent.kt
@@ -1,0 +1,24 @@
+package ui.composables.credentials
+
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.ImageBitmap
+import at.asitplus.wallet.lib.agent.SubjectCredentialStore
+import data.credentials.AgeVerificationCredentialAdapter
+import data.credentials.MobileDrivingLicenceCredentialAdapter
+
+@Composable
+fun AgeVerificationCredentialSummaryCardContent(
+    credential: SubjectCredentialStore.StoreEntry,
+    decodeToBitmap: (ByteArray) -> Result<ImageBitmap>,
+    modifier: Modifier = Modifier,
+) {
+    val credentialAdapter = remember {
+        AgeVerificationCredentialAdapter.createFromStoreEntry(credential)
+    }
+
+    AgeVerificationCredentialAgeDataCard(
+        credentialAdapter = credentialAdapter,
+    )
+}