+++ shared/src/commonMain/kotlin/ui/views/iso/verifier/VerifierPresentationView.kt
@@ -23,6 +23,7 @@ import at.asitplus.valera.resources.error_mdl_driving_privilege_category_expired
 import at.asitplus.valera.resources.error_mdl_driving_privilege_category_not_yet_valid
 import at.asitplus.valera.resources.heading_label_received_data
 import at.asitplus.valera.resources.info_text_credential_status_valid
+import at.asitplus.wallet.ageverification.AgeVerificationScheme
 import at.asitplus.wallet.app.common.decodeImage
 import at.asitplus.wallet.app.common.thirdParty.at.asitplus.wallet.lib.data.iconLabel
 import at.asitplus.wallet.app.common.thirdParty.at.asitplus.wallet.lib.data.uiLabel
@@ -33,6 +34,7 @@ import at.asitplus.wallet.lib.data.IsoDocumentParsed
 import at.asitplus.wallet.mdl.DrivingPrivilege
 import at.asitplus.wallet.mdl.MobileDrivingLicenceDataElements
 import at.asitplus.wallet.mdl.MobileDrivingLicenceScheme
+import data.credentials.AgeVerificationCredentialIsoMdocAdapter
 import data.credentials.EuPidCredentialIsoMdocAdapter
 import data.credentials.HealthIdCredentialIsoMdocAdapter
 import data.credentials.MobileDrivingLicenceCredentialIsoMdocAdapter
@@ -48,6 +50,7 @@ import ui.composables.Logo
 import ui.composables.PersonAttributeDetailCardHeading
 import ui.composables.PersonAttributeDetailCardHeadingIcon
 import ui.composables.buttons.NavigateUpButton
+import ui.composables.credentials.AgeVerificationCredentialViewFromAdapter
 import ui.composables.credentials.CredentialCardLayout
 import ui.composables.credentials.EuPidCredentialViewFromAdapter
 import ui.composables.credentials.HealthIdViewFromAdapter
@@ -158,6 +161,9 @@ fun IsoMdocCredentialView(
                 )
                 is HealthIdScheme -> HealthIdViewFromAdapter(
                     HealthIdCredentialIsoMdocAdapter(namespaces)
+                )
+                is AgeVerificationScheme -> AgeVerificationCredentialViewFromAdapter(
+                    AgeVerificationCredentialIsoMdocAdapter(namespaces)
                 )
                 else -> throw IllegalArgumentException("Unsupported scheme: $scheme")
             }