+++ shared/src/commonMain/kotlin/ui/views/iso/verifier/requests/AVRequests.kt
@@ -13,6 +13,7 @@ import at.asitplus.valera.resources.Res
 import at.asitplus.valera.resources.button_label_check_age
 import at.asitplus.valera.resources.button_label_check_identity
 import at.asitplus.valera.resources.button_label_check_over_age
+import at.asitplus.valera.resources.section_heading_request_av
 import at.asitplus.valera.resources.section_heading_request_pid
 import at.asitplus.valera.resources.text_label_all_attributes
 import at.asitplus.valera.resources.text_label_mandatory_attributes
@@ -22,30 +23,42 @@ import data.document.SelectableRequestType
 import org.jetbrains.compose.resources.stringResource
 
 @Composable
-fun PIDRequests(
+fun AVRequests(
     layoutSpacingModifier: Modifier,
     listSpacingModifier: Modifier,
     onRequestSelected: (SelectableRequest) -> Unit
 ) {
+    val showDropDownAge = remember { mutableStateOf(false) }
+
     val requestItems = listOf(
         RequestItemData(
-            icon = { Icon(imageVector = Icons.Outlined.Person, contentDescription = null) },
-            label = stringResource(Res.string.button_label_check_identity),
-            subLabel = stringResource(Res.string.text_label_mandatory_attributes),
-            onClick = { onRequestSelected(SelectableRequest(SelectableRequestType.PID_MANDATORY)) }
-        ),
-        RequestItemData(
-            icon = { Icon(imageVector = Icons.Outlined.Person, contentDescription = null) },
-            label = stringResource(Res.string.button_label_check_identity),
-            subLabel = stringResource(Res.string.text_label_all_attributes),
-            onClick = { onRequestSelected(SelectableRequest(SelectableRequestType.PID_FULL)) }
+            icon = { Icon(imageVector = Icons.Outlined.Cake, contentDescription = null) },
+            label = stringResource(Res.string.button_label_check_age),
+            onClick = { showDropDownAge.value = !showDropDownAge.value }
         )
     )
 
     RequestSection(
-        title = stringResource(Res.string.section_heading_request_pid),
+        title = stringResource(Res.string.section_heading_request_av),
         layoutSpacingModifier = layoutSpacingModifier,
         listSpacingModifier = listSpacingModifier,
         requestItems = requestItems,
+        extraContent = {
+            if (showDropDownAge.value) {
+                Column {
+                    SelectableAge.valuesList.forEach { age ->
+                        RequestItem(
+                            requestItemData = RequestItemData(
+                                label = stringResource(Res.string.button_label_check_over_age, age),
+                                onClick = {
+                                    onRequestSelected(SelectableRequest(SelectableRequestType.AGE_VERIFICATION, age))
+                                }
+                            ),
+                            modifier = listSpacingModifier
+                        )
+                    }
+                }
+            }
+        }
     )
 }