+++ CHANGELOG.md
@@ -8,6 +8,7 @@
  * Update PIDs to rulebook from October 2025
  * Issuing: Support direct invocation from browser
  * Support URL schemes `haip-vp` and `haip-vci` from latest HAIP draft
+ * Add Age Verification credential
 
 # Release 5.7.3
  * Update to VC-K 5.8.0, fixing optional attributes during presentation